+++ src/resources/views/pages/home.blade.php
@@ -70,6 +70,7 @@
                         <li>Able to change Order statuses with tracking changes in Order model. (issue #138)</li>
                         <li>Added reusable Revision and Revisionable trait for tracking of changes in any model. (issue #138)</li>
                         <li>Added UI for viewing Revision History for Orders.</li>
+                        <li>Allow User to change their password and names (issue #159)</li>
                     </ol>
                     <h4>Enhancements:</h4>
                     <ol>
@@ -115,6 +116,9 @@
                         </li>
                         <li>Improved search engine to be reusable.</li>
                         <li>Payment Statuses can now be edited via config file. See UPGRADE.md for instructions.</li>
+                        <li>Product variant should inherit category from main product (issue #160)</li>
+                        <li>Move printMenu to Html class and partial template (issue #155)</li>
+                        <li>Add field automatically_apply to Coupon model (issue #162)</li>
                     </ol>
                     <h4>Bug fixes:</h4>
                     <ol>
@@ -129,6 +133,7 @@
                         <li>Fixed Typeahead styling messed up for email suggestions. (issue #156)</li>
                         <li>Upload image with same name will now be appended with number. (issue #154)</li>
                         <li>Product variant create/edit form on error doesn't populate previous inputs. (issue #134)</li>
+                        <li>UI: menu unable to scroll on mobile screens (issue #161)</li>
                     </ol>
                     <hr>
                     <h3>Version 0.3.2 and 0.2.2</h3>